+++ packages/bluesky_text/lib/src/entities/entity.dart
@@ -7,6 +7,7 @@
 // 📦 Package imports:
 import 'package:freezed_annotation/freezed_annotation.dart';
 
+import '../atproto.dart';
 import 'byte_indices.dart';
 
 part 'entity.freezed.dart';
@@ -14,6 +15,9 @@ part 'entity.g.dart';
 
 @freezed
 class Entity with _$Entity {
+  // ignore: unused_element
+  const Entity._();
+
   const factory Entity({
     required EntityType type,
     required String value,
@@ -21,6 +25,46 @@ class Entity with _$Entity {
   }) = _Entity;
 
   factory Entity.fromJson(Map<String, Object?> json) => _$EntityFromJson(json);
+
+  /// Returns the facet representation of this entity as JSON.
+  Future<Map<String, dynamic>> toFacet() async {
+    final facet = <String, dynamic>{
+      'index': {
+        'byteStart': indices.start,
+        'byteEnd': indices.end,
+      },
+      'features': []
+    };
+
+    switch (type) {
+      case EntityType.handle:
+        final did = await atproto.identities.findDID(
+          handle: value.substring(1),
+        );
+
+        facet['features'].add({
+          '\$type': 'app.bsky.richtext.facet#mention',
+          'did': did.data.did,
+        });
+
+        break;
+      case EntityType.link:
+        facet['features'].add({
+          '\$type': 'app.bsky.richtext.facet#link',
+          'uri': value,
+        });
+
+        break;
+    }
+
+    return facet;
+  }
+
+  /// Returns true if this entity is handle, otherwise false.
+  bool get isHandle => type == EntityType.handle;
+
+  /// Returns true if this entity is link, otherwise false.
+  bool get isLink => type == EntityType.link;
 }
 
 enum EntityType {